fixed edit bug + common agenda navigation fixes
